requirements.txt
langchain
langchain_community
langchain_groq
python-dotenv
pypdf
langchain_core
dotenv
chromadb
sentence-transformers
chromadb